🧪 Inline ZeroMQ service filepath patching into the test fixture (#7464)
The `_patch_zmq_broker_service_filepaths` helper in `tests/conftest.py` had
a single remaining caller: the `zeromq_broker` fixture in
`tests/brokers/test_zeromq_broker.py`. Inline its logic into that fixture
and remove the shared helper, along with the now-unused `patch` and
`contextmanager` imports, from `conftest.py`.
